refactor: move lockfile discovery/loading into Packages.load
The filesystem work — walking up to npm's lockfile (the monorepo root in a
workspace), reading it and any linked-dep child lockfiles, computing the
cwd prefix — is a Packages concern, not a Nudeps one. Extract it into a
static Packages.load(cwd, { warn }) factory; the constructor stays a pure,
filesystem-free parser (as the tests rely on).

Nudeps.packages now just delegates, the standalone modules getter is gone,
and prefix lives on the Packages instance (map.js reads packages.prefix).
